# back/central-reserve/shared/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.CreatePermissionRequest:
    properties:
      action_id:
        description: ID de la action
        example: 1
        type: integer
      business_type_id:
        example: 11
        type: integer
      code:
        description: Opcional, se genera automáticamente si no se proporciona
        example: horizontalproperty_createuser
        type: string
      description:
        example: Permite crear nuevos usuarios en el sistema
        type: string
      name:
        example: Crear usuario
        type: string
      resource_id:
        description: ID del resource
        example: 1
        type: integer
      scope_id:
        example: 1
        type: integer
    required:
    - action_id
    - name
    - resource_id
    - scope_id
    type: object
...
  request.CreateResourceRequest:
    description: Solicitud para crear un nuevo recurso en el sistema
    properties:
      business_type_id:
        example: 11
        type: integer
      description:
        example: Gestión de usuarios del sistema
        type: string
      name:
        example: users
        type: string
    required:
    - name
    type: object
...
  request.UpdateResourceRequest:
    description: Solicitud para actualizar un recurso existente
    properties:
      business_type_id:
        example: 11
        type: integer
      description:
        example: Gestión de usuarios del sistema
        type: string
      name:
        example: users
        type: string
    required:
    - name
    type: object
...
  response.PermissionErrorResponse:
    properties:
      error:
        example: Error al procesar la solicitud
        type: string
    type: object
  response.PermissionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/services_auth_internal_infra_primary_controllers_permissionhandler_response.PermissionResponse'
        type: array
      success:
        example: true
        type: boolean
      total:
        example: 10
        type: integer
    type: object
  response.PermissionMessageResponse:
    properties:
      message:
        example: Permiso creado exitosamente
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  services_auth_internal_infra_primary_controllers_permissionhandler_response.PermissionResponse:
    properties:
      action:
        example: create
        type: string
      business_type_id:
        example: 11
        type: integer
      business_type_name:
        example: Propiedad Horizontal
        type: string
      code:
        example: users:create
        type: string
      description:
        example: Permite crear nuevos usuarios en el sistema
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Crear usuarios
        type: string
      resource:
        example: users
        type: string
      scope_code:
        example: system
        type: string
      scope_id:
        example: 1
        type: integer
      scope_name:
        example: Sistema
        type: string
    type: object
...
host: localhost:3050
info:
  contact:
    email: backend@example.com
    name: Equipo de Backend
  description: Servicio REST para la gestión de reservas multi-restaurante.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://ejemplo.com/terminos
  title: Restaurant Reservation API
  version: "1.0"
paths:
...
  /permissions:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de permisos del sistema con filtros opcionales
        por tipo de business, nombre y scope
      parameters:
      - description: Filtrar por tipo de business (incluye genéricos)
        in: query
        name: business_type_id
        type: integer
      - description: Filtrar por nombre de permiso (búsqueda parcial)
        in: query
        name: name
        type: string
      - description: Filtrar por ID de scope
        in: query
        name: scope_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de permisos obtenida exitosamente
          schema:
            $ref: '#/definitions/response.PermissionListResponse'
        "401":
          description: Token de acceso requerido
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener todos los permisos
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: Crea un nuevo permiso en el sistema
      parameters:
      - description: Datos del permiso a crear
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/request.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permiso creado exitosamente
          schema:
            $ref: '#/definitions/response.PermissionMessageResponse'
        "400":
          description: Datos de entrada inválidos
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
        "401":
          description: Token de acceso requerido
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
        "409":
          description: Permiso con código o nombre duplicado
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.PermissionErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear nuevo permiso
      tags:
      - Permissions
...
  /resources:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de recursos del sistema con opciones
        de filtrado y ordenamiento
      parameters:
      - default: 1
        description: Número de página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Tamaño de página
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Filtrar por nombre (búsqueda parcial)
        in: query
        name: name
        type: string
      - description: Filtrar por descripción (búsqueda parcial)
        in: query
        name: description
        type: string
      - description: Filtrar por tipo de business (incluye genéricos)
        in: query
        name: business_type_id
        type: integer
      - description: Campo para ordenar
        enum:
        - name
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Orden
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de recursos obtenida exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parámetros de consulta inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener recursos
      tags:
      - Resources
    post:
      consumes:
      - application/json
      description: Crea un nuevo recurso en el sistema con nombre y descripción únicos
      parameters:
      - description: Datos del recurso a crear
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Recurso creado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Recurso ya existe
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear recurso
      tags:
      - Resources
  /resources/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina permanentemente un recurso del sistema por su ID único
        con eliminación en cascada
      parameters:
      - description: ID del recurso
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recurso eliminado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID de recurso inválido
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recurso no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar recurso
      tags:
      - Resources
    get:
      consumes:
      - application/json
      description: Obtiene un recurso específico del sistema por su ID único
      parameters:
      - description: ID del recurso
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recurso obtenido exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID de recurso inválido
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recurso no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener recurso por ID
      tags:
      - Resources
    put:
      consumes:
      - application/json
      description: Actualiza un recurso existente en el sistema
      parameters:
      - description: ID del recurso
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Datos del recurso a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Recurso actualizado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recurso no encontrado
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflicto con recurso existente
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar recurso
      tags:
      - Resources
...
securityDefinitions:
  BearerAuth:
    description: Token de business JWT con el prefijo **Bearer** (para todos los endpoints)
    in: header
    name: Authorization
    type: apiKey
  BusinessTokenAuth:
    description: Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"